+++ internal/handlers/products.go
